backend/requirements.txt
flask
flask-cors
groq
pdfplumber
python-docx
python-pptx
gunicorn
